Centralize contracts into a single package
Instead of having contracts splayed across the source tree, we now
have a single `contracts` package sitting in `pkg`.
