+++ cli/azd/pkg/contracts/console_message.go
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package contracts
+
+type ConsoleMessage struct {
+	Message string `json:"message"`
+}
