+++ cli/azd/pkg/contracts/login.go
@@ -1,5 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+
 package contracts
 
 import "time"
