+++ cli/azd/pkg/contracts/show.go
